substratus/eval/requirements.txt
datasets
weaviate-client

torch
xformers
transformers
peft
accelerate
huggingface_hub
bitsandbytes
protobuf

scipy
einops
